Move "patterns" folder -> "docsUtil" folder
