+++ src/app/components/index.tsx
@@ -4,7 +4,7 @@ export * from './layout';
 export * from './navigation';
 export * from './resources';
 export * from './InfoCard';
-export * from './patterns';
+export * from './docsUtil';
 export * from './contributors';
 export * from './faq';
 export * from './typography';
